Adding support in portal service to handle query strings
