+++ AzureFunctions.AngularClient/src/app/shared/Utilities/url.ts
@@ -1,0 +1,21 @@
+export class Url{
+    public static getParameterByName(url, name) {
+        if (url === null) {
+            url = window.location.href;
+        }
+
+        name = name.replace(/[\[\]]/g, '\\$&');
+        let regex = new RegExp('[?&]' + name + '(=([^&#]*)|&|#|$)');
+        let results = regex.exec(url);
+
+        if (!results) {
+            return null;
+        }
+
+        if (!results[2]) {
+            return '';
+        }
+
+        return decodeURIComponent(results[2].replace(/\+/g, ' '));
+    }
+}
